requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
executing==2.2.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.69.1
griffe==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logfire==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.30.0
numpy==2.2.3
openai==1.66.3
openai-agents==0.0.4
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wrapt==1.17.2
zipp==3.21.0
fastapi==0.115.6
uvicorn==0.32.1